libretro: drive UBSan to zero findings; clean up every build warning
Sanitizer work first, with a process disclosure it surfaced: the sanitizer
runs recorded in the LZMA, Tremor and cleanup commits were vacuous -- the
builds lacked a preceding make clean, so make SANITIZER=... found fresh
release objects and relinked uninstrumented binaries (this Makefile has no
header dependency tracking, and nm showed zero __asan/__ubsan symbols).
The claims in those messages happened to be re-validated by the genuine
clean instrumented builds behind this commit (lifecycle exit 0, zero
findings, zero leaks), but the record should say plainly that the original
runs proved nothing. Every sanitizer figure below is from a verified
instrumented build (nm -D | grep -c __ubsan checked non-zero).

A genuine full inventory showed 122 UBSan findings. Now zero:

- ~110 were one root: ClassDef::InitializeActorClass creates actor default
  instances as raw M_Malloc + parent-memcpy blobs that never run a
  constructor, because the DObject constructor enrols every object in the
  GC root list and defaults must not participate in GC. Every DECORATE
  property handler then downcasts and writes fields on memory with no
  dynamic type -- benign by construction (layouts match), but exactly what
  -fsanitize=vptr flags. The sanitizer block now carves out vptr with a
  comment documenting the pattern; every other UBSan check stays on.
- DECLARE_POINTER built member offsets as (size_t)&((ThisClass*)1)->ptr-1,
  a member-operator call on a misaligned non-object at static init, firing
  once per pointer declaration across a dozen TUs. It now uses offsetof,
  which is conditionally-supported on these non-standard-layout types and
  well-defined on every toolchain this core builds with (the matching
  C++-only -Wno-invalid-offsetof is scoped to CXXFLAGS, since the flag is
  not valid for C).
- Real bugs fixed: StateDefinition and Scanner::ParserState left POD
  members uninitialized and compiler-generated copies read them (invalid
  bool loads of values like 26 and 201); w_wad hashed lump tables through
  memset(NULL, ...) before any file was loaded, and read from a null cache
  with length zero; thingdef qsorted an empty symbol table through a null
  base; InstallStates formed &frameList[0] and one-past-end references on
  state-less actors, and the GetState / IsStateOwner header inlines did
  the same during label resolution (both are now null-safe); two texture
  probes multiplied header-supplied dimensions in int (signed overflow on
  hostile input, now int64_t).

Warning cleanup, to a zero-warning clean build (from 90+):

- id_sd_adlib's post-render shrink computed the new cursor from the freed
  pointer after realloc (-Wuse-after-free, formally undefined): the
  rendered count is now taken before the call.
- TSharedPtr's Dereference could not prove to GCC that the static null
  sentinel is never deleted (-Wfree-nonheap-object); the invariant is now
  an explicit local guard.
- FMultiPatchTexture bounded its parts memcpy (GCC saw a signed size range
  through the int member; the copy now keys off the unsigned source size
  and skips the empty case, which also formed &parts[0] on nothing).
- Mix_Chunk_Digital's initializer lists are in declaration order (the
  member block starts with busyTicksOverride, so the old lists reordered
  under -Wreorder in every TU including state_machine.h -- 26 sites).
- wl_floorceiling's texture metrics are zero-initialized; GCC cannot prove
  the guarded uses dominated (-Wmaybe-uninitialized) and a deterministic
  zero is strictly safer.
- Dead code removed (numfiles, lastgamemusicoffset, iyscale, the unused
  USL_XORICursor, a dead newobj binding whose Spawn call is kept for its
  side effects); intentionally-unused ACTION_PARAM decodes are (void)'d at
  each site rather than globally silenced.

Verification: audio AND video stream hashes byte-identical to the previous
build on the 35 fps attract and scripted 120 fps gameplay runs (repeated),
load time unchanged; instrumented lifecycle exit 0 with zero UBSan
findings, zero ASan errors, zero leaks, no suppressions beyond the
documented vptr carve-out.

+++ src/textures/rottflattexture.cpp
@@ -81,7 +81,7 @@ static bool CheckIfRottFlat(FileReader &file)
 
 	uint16_t Width = LittleShort(header[0]);
 	uint16_t Height = LittleShort(header[1]);
-	if(file.GetLength() == Width*Height+8)
+	if(file.GetLength() == (int64_t)Width*Height+8)
 		return true;
 	return false;
 }
