+++ src/textures/wolfrawtexture.cpp
@@ -83,12 +83,12 @@ static bool CheckIfWolfRaw(FileReader &file)
 
 	uint16_t Width = LittleShort(header[0]);
 	uint16_t Height = LittleShort(header[1]);
-	if(file.GetLength() == Width*Height+4) // Raw page
+	if(file.GetLength() == (int64_t)Width*Height+4) // Raw page
 		return true;
 
 	Width = BigShort(header[0]);
 	Height = BigShort(header[1]);
-	if(file.GetLength() == Width*Height+4) // Mac raw
+	if(file.GetLength() == (int64_t)Width*Height+4) // Mac raw
 		return true;
 	return false;
 }
